1. Модульные тесты. [#159435]

+++ src/test/java/ru/job4j/converter/ConverterTest.java
@@ -17,7 +17,7 @@ public class ConverterTest {
     @Test
     public void whenConvert140RblThen2dot3333Dlr() {
         float in = 140;
-        float expected = (float) 2.3333;
+        float expected = 2.3333F;
         float out = Converter.rubleToDollar(in);
         float eps = 0.0001f;
         Assert.assertEquals(expected, out, eps);
